pyproject.toml
[tool.setuptools.packages.find]
where = ["."]
include = ["model", "route", "services", "assistants"]
exclude = ["frontend", "saved_files"]


[project]
name = "pre-sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.124.4",
    "langsmith[openai-agents]>=0.5.0",
    "openai-agents[sqlalchemy]>=0.6.2",
    "psycopg>=3.3.2",
    "py-dotenv>=0.1",
    "pymongo[srv]>=4.15.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "weaviate-client>=4.19.2",
]
